Fix linter comments
