+++ src/mapper/pkg/resolvers/schema.resolvers.go
@@ -21,54 +21,10 @@ import (
 	"golang.org/x/exp/slices"
 )
 
-func (r *mutationResolver) ResetCapture(_ context.Context) (bool, error) {
+func (r *mutationResolver) ResetCapture(ctx context.Context) (bool, error) {
 	logrus.Info("Resetting stored intents")
 	r.intentsHolder.Reset()
 	return true, nil
-}
-
-func (r *mutationResolver) discoverSrcIdentity(ctx context.Context, src model.RecordedDestinationsForSrc) *model.OtterizeServiceIdentity {
-	srcPod, err := r.kubeFinder.ResolveIpToPod(ctx, src.SrcIP)
-	if err != nil {
-		if errors.Is(err, kubefinder.ErrFoundMoreThanOnePod) {
-			logrus.WithError(err).Debugf("Ip %s belongs to more than one pod, ignoring", src.SrcIP)
-		} else {
-			logrus.WithError(err).Debugf("Could not resolve %s to pod", src.SrcIP)
-		}
-		return nil
-	}
-	if src.SrcHostname != "" && srcPod.Name != src.SrcHostname {
-		// This could mean a new pod is reusing the same IP
-		// TODO: Use the captured hostname to actually find the relevant pod (instead of the IP that might no longer exist or be reused)
-		logrus.Warnf("Found pod %s (by ip %s) doesn't match captured hostname %s, ignoring", srcPod.Name, src.SrcIP, src.SrcHostname)
-		return nil
-	}
-
-	if srcPod.DeletionTimestamp != nil {
-		logrus.Debugf("Pod %s is being deleted, ignoring", srcPod.Name)
-		return nil
-	}
-
-	srcService, err := r.serviceIdResolver.ResolvePodToServiceIdentity(ctx, srcPod)
-	if err != nil {
-		logrus.WithError(err).Debugf("Could not resolve pod %s to identity", srcPod.Name)
-		return nil
-	}
-	filteredDestinations := make([]model.Destination, 0)
-	for _, dest := range src.Destinations {
-		if srcPod.CreationTimestamp.After(dest.LastSeen) {
-			logrus.Debugf("Pod %s was created after capture time %s, ignoring", srcPod.Name, dest.LastSeen)
-			continue
-		}
-		filteredDestinations = append(filteredDestinations, dest)
-	}
-
-	srcSvcIdentity := model.OtterizeServiceIdentity{Name: srcService.Name, Namespace: srcPod.Namespace, Labels: podLabelsToOtterizeLabels(srcPod)}
-	if srcService.OwnerObject != nil {
-		srcSvcIdentity.PodOwnerKind = model.GroupVersionKindFromKubeGVK(srcService.OwnerObject.GetObjectKind().GroupVersionKind())
-	}
-
-	return &srcSvcIdentity
 }
 
 func (r *mutationResolver) ReportCaptureResults(ctx context.Context, results model.CaptureResults) (bool, error) {
@@ -305,7 +261,7 @@ func (r *mutationResolver) ReportIstioConnectionResults(ctx context.Context, res
 	return true, nil
 }
 
-func (r *queryResolver) ServiceIntents(_ context.Context, namespaces []string, includeLabels []string, includeAllLabels *bool) ([]model.ServiceIntents, error) {
+func (r *queryResolver) ServiceIntents(ctx context.Context, namespaces []string, includeLabels []string, includeAllLabels *bool) ([]model.ServiceIntents, error) {
 	shouldIncludeAllLabels := false
 	if includeAllLabels != nil && *includeAllLabels {
 		shouldIncludeAllLabels = true
@@ -330,7 +286,7 @@ func (r *queryResolver) ServiceIntents(_ context.Context, namespaces []string, i
 	return intentsBySource, nil
 }
 
-func (r *queryResolver) Intents(_ context.Context, namespaces []string, includeLabels []string, excludeServiceWithLabels []string, includeAllLabels *bool) ([]model.Intent, error) {
+func (r *queryResolver) Intents(ctx context.Context, namespaces []string, includeLabels []string, excludeServiceWithLabels []string, includeAllLabels *bool) ([]model.Intent, error) {
 	shouldIncludeAllLabels := false
 	if includeAllLabels != nil && *includeAllLabels {
 		shouldIncludeAllLabels = true
