+++ src/sniffer/pkg/ipresolver/procfs_resolver.go
@@ -4,109 +4,7 @@ import (
 	"errors"
 	"github.com/otterize/network-mapper/src/sniffer/pkg/utils"
 	"github.com/sirupsen/logrus"
-	"sync"
 )
-
-type ProcessMonitor struct {
-	processes  map[int64]interface{}
-	onProcNew  utils.ProcessScanCallback
-	onProcExit utils.ProcessScanCallback
-	pollEvent  chan bool // TODO: Replace with sync.Cond if supporting multiple callers is needed
-	pollLock   sync.Mutex
-	done       chan bool
-}
-
-func NewProcessMonitor(onProcNew, onProcExit utils.ProcessScanCallback) *ProcessMonitor {
-	return &ProcessMonitor{
-		processes:  make(map[int64]interface{}),
-		onProcNew:  onProcNew,
-		onProcExit: onProcExit,
-		pollEvent:  make(chan bool, 1),
-		done:       nil,
-	}
-}
-
-//func (pm *ProcessMonitor) Start(intervals ...int) {
-//	var interval int
-//	if len(intervals) > 0 {
-//		interval = intervals[0] * 1000
-//	} else {
-//		interval = 500 // default interval
-//	}
-//
-//	pm.done = make(chan bool)
-//
-//	go func() {
-//		for {
-//			select {
-//			case <-pm.done:
-//				return
-//			default:
-//				pm.pollLock.Lock()
-//				err := pm.poll()
-//
-//				select { // Raise event unless channel already full
-//				case pm.pollEvent <- true:
-//				default:
-//				}
-//				pm.pollLock.Unlock()
-//
-//				if err != nil {
-//					logrus.Errorf("ProcessMonitor: poll failed: %v", err)
-//				}
-//				time.Sleep(time.Duration(interval) * time.Millisecond)
-//			}
-//		}
-//	}()
-//}
-
-//func (pm *ProcessMonitor) Stop() {
-//	if pm.done != nil {
-//		pm.done <- true
-//	}
-//}
-//
-//func (pm *ProcessMonitor) WaitForNextPoll() {
-//	pm.pollLock.Lock() // Lock ensures poll hasn't already started, we're waiting for entirely new refresh
-//	select {
-//	case <-pm.pollEvent: // Reset event if it's already set
-//	default:
-//	}
-//	pm.pollLock.Unlock()
-//
-//	<-pm.pollEvent // Wait for next poll to complete
-//}
-
-func (pm *ProcessMonitor) poll() error {
-	oldProcesses := make(map[int64]bool)
-	for pid := range pm.processes {
-		oldProcesses[pid] = false
-	}
-
-	err := utils.ScanProcDirProcesses(func(pid int64, pDir string) {
-		if _, ok := pm.processes[pid]; !ok {
-			// New process
-			pm.onProcNew(pid, pDir)
-			pm.processes[pid] = nil
-		} else {
-			// Existing process
-			oldProcesses[pid] = true
-		}
-	})
-	if err != nil {
-		return err
-	}
-
-	for pid := range oldProcesses {
-		if !oldProcesses[pid] {
-			pm.onProcExit(pid, "")
-			// Process no longer exists
-			delete(pm.processes, pid)
-		}
-	}
-
-	return nil
-}
 
 type ProcFSIPResolverEntry struct {
 	IPAddr          string
@@ -127,13 +25,9 @@ func NewProcFSIPResolver() *ProcFSIPResolver {
 		byPid:   make(map[int64]*ProcFSIPResolverEntry),
 	}
 	r.monitor = NewProcessMonitor(r.onProcessNew, r.onProcessExit)
-	//r.monitor.Start()
+
 	return &r
 }
-
-//func (r *ProcFSIPResolver) Stop() {
-//	r.monitor.Stop()
-//}
 
 func (r *ProcFSIPResolver) ResolveIP(ipaddr string) (hostname string, err error) {
 	if hostInfo, ok := r.byAddr[ipaddr]; ok {
@@ -143,12 +37,8 @@ func (r *ProcFSIPResolver) ResolveIP(ipaddr string) (hostname string, err error)
 }
 
 func (r *ProcFSIPResolver) Refresh() error {
-	return r.monitor.poll()
+	return r.monitor.Poll()
 }
-
-//func (r *ProcFSIPResolver) WaitForNextRefresh() {
-//	r.monitor.WaitForNextPoll()
-//}
 
 func (r *ProcFSIPResolver) onProcessNew(pid int64, pDir string) {
 	hostname, err := utils.ExtractProcessHostname(pDir)
